Fixed Aran Movement Bug (when attacking) Fixed Body Pressure + Drain bug Fixed some guild bugs Fixed Taunt and Ninja Ambush Added Miu Miu store (run MiuMiu.sql)
git-svn-id: https://svn6.assembla.com/svn/MoopleDEV@115 59296cf3-ab24-4890-a14e-c0dede3ece5d

+++ src/server/movement/AbsoluteLifeMovement.java
@@ -28,7 +28,7 @@ public class AbsoluteLifeMovement extends AbstractLifeMovement {
     private Point pixelsPerSecond;
     private int unk;
 
-    public AbsoluteLifeMovement(int type, Point position, int duration, int newstate) {
+    public AbsoluteLifeMovement(byte type, Point position, int duration, byte newstate) {
         super(type, position, duration, newstate);
     }
