+++ src/server/movement/LifeMovement.java
@@ -25,7 +25,7 @@ import java.awt.Point;
 
 public interface LifeMovement extends LifeMovementFragment {
     Point getPosition();
-    int getNewstate();
+    byte getNewstate();
     int getDuration();
-    int getType();
+    byte getType();
 }
